All options overwritable per wiki page
# Changes overview

## syntax/theme.php

### Parse options as query string

Allow the old syntax for only the theme name or the the new syntax for
more options as a query string like

```
~~REVEAL theme=sky&transition=convex&controls=1~~
```

### Put all options except `open_in_new_window` to the export URL

## renderer.php

### Merge URL params into plugin conf

Changing parameters direct in the URL is only working, when the page is
not cached (`~~NOCACHE~~`).

The good thing with `~~NOCACHE~~` is, that you are able to change the
layout on the fly by only changing the URL. If you encounter any
problems during a presentation e.g. bad contrast, then you are able to
change this without editing the wiki page.

## example_presentation.dokuwiki

### Example parameters, new List page and cleanup